Fill in wireframe
Quick layout of wireframe elements, without styling, in HTML. Includes a block
of CSS to mark the blocks as I move them around with various colors of borders.
Styling of the elements will be done after the layout is on target.

Convert a few of the provided pieces to my standard process, including the font
stylesheet in the style section rather than as a link in HTML head.

Drop the double-up of HTML from importing the provided code into my boilerplate.

Add the HTML elements of my wireframe and move the provided text into it.

Signed-off-by: Ronald Lamoreaux <upskill@chindraba.work>